Lucene now properly re-indexes upon refresh. Moved the SearchResult entity to the DTO
